silence a couple conversion warnings

+++ src/ui/Suite.cpp
@@ -62,8 +62,8 @@ void Suite::resize()
 	mGraph->setNeedsLayout();
 
 	const float width = 120; // TODO: calculate widest segment
-	const float height = 22 * mSelector->getSegmentLabels().size();
-	const float windowWidth = app::getWindowWidth();
+	const float height = float( 22 * mSelector->getSegmentLabels().size() );
+	const float windowWidth = (float)app::getWindowWidth();
 	mSelector->setBounds( Rectf( windowWidth - width - PADDING, PADDING, windowWidth - PADDING, height + PADDING ) );
 }
